Fixed the list button after moving the map slice
